add error message

+++ Collection/views.py
@@ -8,7 +8,7 @@ from django.views.decorators.http import require_http_methods
 
 from common.deco import user_logged_in
 from common.functool import checkParameter, getFirstProductImageId, getReqUser
-from common.restool import resBadRequest, resForbidden, resInvalidPara, resMissingPara, resOk, resReturn
+from common.restool import resBadRequest, resForbidden, resInvalidPara, resMissingPara, resOk, resReturn, resUnauthorized
 
 # Create your views here.
 
@@ -19,16 +19,18 @@ def createCollection(request):
         return resMissingPara(['productId'])
 
     data = json.loads(request.body)
-    productId=data['productId']
+    productId = data['productId']
 
     if productIdValidation(productId):
         user = getReqUser(request)
+        if user is None:
+            return resUnauthorized("用户未登录")
         product = Product.objects.get(id=productId)
         if not Collection.objects.filter(user=user, product=product).exists():
             collection = Collection.objects.create(user=user, product=product)
             return resReturn({'collectionId' : collection.id})
         else:
-            return resBadRequest("Collection exists.")
+            return resBadRequest("该商品已收藏过")
     else:
         return resInvalidPara(["productId"])
     
@@ -37,10 +39,12 @@ def createCollection(request):
 @user_logged_in
 def deleteCollection(request,id):
     user = getReqUser(request)
+    if user is None:
+        return resUnauthorized("用户未登录")
     collection = get_object_or_404(Collection, id=id)
 
     if user.id != collection.user.id:
-        return resForbidden()
+        return resForbidden("不可删除其他用户的收藏")
 
     collection.delete()
     return resOk()
@@ -49,5 +53,7 @@ def deleteCollection(request,id):
 @user_logged_in
 def collectionList(request):
     user = getReqUser(request)
+    if user is None:
+        return resUnauthorized("用户未登录")
     collections = Collection.objects.filter(user=user)
     return resReturn({"result" : [{'product' : c.product.body(), 'image' : getFirstProductImageId(c.product)} for c in collections]})
